Convert standalone example into a test
